+++ src/Sentry/IsExternalInit.cs
@@ -1,0 +1,6 @@
+﻿#if !NET5_0_OR_GREATER
+namespace System.Runtime.CompilerServices
+{
+    internal static class IsExternalInit {}
+}
+#endif