+++ test/Sentry.Tests/Internals/Http/CachingTransportTests.cs
@@ -337,6 +337,7 @@ public class CachingTransportTests
         innerTransport
             .SendEnvelopeAsync(Arg.Any<Envelope>(), Arg.Any<CancellationToken>())
             .Returns(_ =>
+                // ReSharper disable once AccessToModifiedClosure
                 isFailing
                     ? Task.FromException(new InvalidOperationException())
                     : Task.CompletedTask);
@@ -360,6 +361,128 @@ public class CachingTransportTests
         // Assert
         // (0 envelopes retried)
         _ = innerTransport.Received(0).SendEnvelopeAsync(Arg.Any<Envelope>(), Arg.Any<CancellationToken>());
+    }
+
+    [Fact]
+    public async Task RecordsDiscardedEventOnNonTransientExceptions()
+    {
+        // Arrange
+        using var cacheDirectory = new TempDirectory();
+        var options = new SentryOptions
+        {
+            Dsn = DsnSamples.ValidDsnWithoutSecret,
+            DiagnosticLogger = _logger,
+            Debug = true,
+            CacheDirectoryPath = cacheDirectory.Path,
+            ClientReportRecorder = Substitute.For<IClientReportRecorder>()
+        };
+
+
+        var innerTransport = Substitute.For<ITransport>();
+        innerTransport
+            .SendEnvelopeAsync(Arg.Any<Envelope>(), Arg.Any<CancellationToken>())
+            .Returns(_ => Task.FromException(new InvalidOperationException()));
+
+        await using var transport = CachingTransport.Create(innerTransport, options, startWorker: false);
+
+        // Act
+        using var envelope = Envelope.FromEvent(new SentryEvent());
+        await transport.SendEnvelopeAsync(envelope);
+        await transport.FlushAsync();
+
+        // Test that we recorded the discarded event
+        options.ClientReportRecorder.Received(1)
+            .RecordDiscardedEvent(DiscardReason.CacheOverflow, DataCategory.Error);
+    }
+
+    [Fact]
+    public async Task RoundtripsClientReports()
+    {
+        // Arrange
+        using var cacheDirectory = new TempDirectory();
+        var options = new SentryOptions
+        {
+            Dsn = DsnSamples.ValidDsnWithoutSecret,
+            DiagnosticLogger = _logger,
+            Debug = true,
+            CacheDirectoryPath = cacheDirectory.Path
+        };
+
+        var timestamp = DateTimeOffset.UtcNow;
+        var clock = Substitute.For<ISystemClock>();
+        clock.GetUtcNow().Returns(timestamp);
+        var recorder = new ClientReportRecorder(options, clock);
+        options.ClientReportRecorder = recorder;
+
+        var innerTransport = new FakeTransportWithRecorder(recorder);
+        await using var transport = CachingTransport.Create(innerTransport, options, startWorker: false);
+
+        // Act
+        recorder.RecordDiscardedEvent(DiscardReason.BeforeSend, DataCategory.Error);
+        using var envelope = Envelope.FromEvent(new SentryEvent());
+        await transport.SendEnvelopeAsync(envelope); // should internally generate a client report and write to disk
+        var interimClientReport = recorder.GenerateClientReport(); // should be null
+        await transport.FlushAsync(); // will read from disk and should send that client report
+
+        // Test that the interim report was null
+        Assert.Null(interimClientReport);
+
+        // Test that we actually sent the discarded event
+        var clientReport = new ClientReport(timestamp,
+            new Dictionary<DiscardReasonWithCategory, int>
+            {
+                {DiscardReason.BeforeSend.WithCategory(DataCategory.Error), 1}
+            }
+        );
+        var expected = await EnvelopeItem.FromClientReport(clientReport).Payload.SerializeToStringAsync(_logger);
+
+        var envelopeSent = innerTransport.GetSentEnvelopes().Single();
+        var envelopeItem = envelopeSent.Items.Single(x => x.TryGetType() == "client_report");
+        var actual = await envelopeItem.Payload.SerializeToStringAsync(_logger);
+        Assert.Equal(expected, actual);
+    }
+
+    [Fact]
+    public async Task RestoresDiscardedEventCounts()
+    {
+        // Arrange
+        using var cacheDirectory = new TempDirectory();
+        var options = new SentryOptions
+        {
+            Dsn = DsnSamples.ValidDsnWithoutSecret,
+            DiagnosticLogger = _logger,
+            Debug = true,
+            CacheDirectoryPath = cacheDirectory.Path
+        };
+
+        var recorder = (ClientReportRecorder) options.ClientReportRecorder;
+        var innerTransport = new FakeTransportWithRecorder(recorder);
+        await using var transport = CachingTransport.Create(innerTransport, options,
+            startWorker: false, failStorage: true);
+
+        // some arbitrary discarded events ahead of time
+        recorder.RecordDiscardedEvent(DiscardReason.BeforeSend, DataCategory.Attachment);
+        recorder.RecordDiscardedEvent(DiscardReason.EventProcessor, DataCategory.Error);
+        recorder.RecordDiscardedEvent(DiscardReason.EventProcessor, DataCategory.Error);
+        recorder.RecordDiscardedEvent(DiscardReason.QueueOverflow, DataCategory.Security);
+        recorder.RecordDiscardedEvent(DiscardReason.QueueOverflow, DataCategory.Security);
+        recorder.RecordDiscardedEvent(DiscardReason.QueueOverflow, DataCategory.Security);
+
+        // Act
+        await Assert.ThrowsAnyAsync<Exception>(async () =>
+        {
+            using var envelope = Envelope.FromEvent(new SentryEvent());
+            await transport.SendEnvelopeAsync(envelope); // will fail, since we set failStorage to true
+        });
+
+        // Assert
+        recorder.DiscardedEvents.Should().BeEquivalentTo(new Dictionary<DiscardReasonWithCategory, int>
+        {
+            // These are the original items recorded.  They should still be there.
+            {DiscardReason.BeforeSend.WithCategory(DataCategory.Attachment), 1},
+            {DiscardReason.EventProcessor.WithCategory(DataCategory.Error), 2},
+            {DiscardReason.QueueOverflow.WithCategory(DataCategory.Security), 3}
+        });
     }
 
     public static IEnumerable<object[]> NetworkTestData =>