+++ test/Sentry.Tests/Internals/Http/HttpTransportTests.cs
@@ -279,6 +279,7 @@ public class HttpTransportTests
             {
                 Dsn = DsnSamples.ValidDsnWithSecret,
                 DiagnosticLogger = new TraceDiagnosticLogger(SentryLevel.Debug),
+                SendClientReports = false,
                 Debug = true
             },
             new HttpClient(httpHandler));
@@ -326,6 +327,170 @@ public class HttpTransportTests
 
         // Assert
         actualEnvelopeSerialized.Should().BeEquivalentTo(expectedEnvelopeSerialized);
+    }
+
+    [Fact]
+    public async Task SendEnvelopeAsync_RateLimited_CountsDiscardedEventsCorrectly()
+    {
+        // Arrange
+        using var httpHandler = new RecordingHttpMessageHandler(
+            new FakeHttpMessageHandler(
+                () => SentryResponses.GetRateLimitResponse("1234:event, 897:transaction")
+            ));
+
+        var timestamp = DateTimeOffset.UtcNow;
+        var fakeClock = Substitute.For<ISystemClock>();
+        fakeClock.GetUtcNow().Returns(timestamp);
+
+        var options = new SentryOptions
+        {
+            Dsn = DsnSamples.ValidDsnWithSecret,
+            DiagnosticLogger = new TraceDiagnosticLogger(SentryLevel.Debug),
+            SendClientReports = true,
+            Debug = true
+        };
+
+        var recorder = new ClientReportRecorder(options, fakeClock);
+        options.ClientReportRecorder = recorder;
+
+        var httpTransport = new HttpTransport(
+            options,
+            new HttpClient(httpHandler),
+            clock: fakeClock
+        );
+
+        // First request always goes through
+        await httpTransport.SendEnvelopeAsync(Envelope.FromEvent(new SentryEvent()));
+
+        var envelope = new Envelope(
+            new Dictionary<string, object>(),
+            new[]
+            {
+                // Should be dropped
+                new EnvelopeItem(
+                    new Dictionary<string, object> {["type"] = "event"},
+                    new EmptySerializable()),
+                new EnvelopeItem(
+                    new Dictionary<string, object> {["type"] = "event"},
+                    new EmptySerializable()),
+                new EnvelopeItem(
+                    new Dictionary<string, object> {["type"] = "transaction"},
+                    new EmptySerializable()),
+
+                // Should stay
+                new EnvelopeItem(
+                    new Dictionary<string, object> {["type"] = "other"},
+                    new EmptySerializable())
+            });
+
+        // The client report should contain rate limit discards only.
+        var expectedClientReport =
+            new ClientReport(timestamp,
+                new Dictionary<DiscardReasonWithCategory, int>
+                {
+                    {DiscardReason.RateLimitBackoff.WithCategory(DataCategory.Error), 2},
+                    {DiscardReason.RateLimitBackoff.WithCategory(DataCategory.Transaction), 1}
+                });
+
+        var expectedEnvelope = new Envelope(
+            new Dictionary<string, object>(),
+            new[]
+            {
+                new EnvelopeItem(
+                    new Dictionary<string, object> {["type"] = "other"},
+                    new EmptySerializable()),
+                EnvelopeItem.FromClientReport(expectedClientReport)
+            });
+
+        var expectedEnvelopeSerialized = await expectedEnvelope.SerializeToStringAsync(new TraceDiagnosticLogger(SentryLevel.Debug));
+
+        // Act
+        await httpTransport.SendEnvelopeAsync(envelope);
+
+        var lastRequest = httpHandler.GetRequests().Last();
+        var actualEnvelopeSerialized = await lastRequest.Content.ReadAsStringAsync();
+
+        // Assert
+        actualEnvelopeSerialized.Should().BeEquivalentTo(expectedEnvelopeSerialized);
+    }
+
+    [Fact]
+    public async Task SendEnvelopeAsync_Fails_RestoresDiscardedEventCounts()
+    {
+        // Arrange
+        using var httpHandler = new RecordingHttpMessageHandler(
+            new FakeHttpMessageHandler(
+                () => new HttpResponseMessage(HttpStatusCode.InternalServerError)));
+
+        var options = new SentryOptions
+        {
+            Dsn = DsnSamples.ValidDsnWithSecret,
+            DiagnosticLogger = new TraceDiagnosticLogger(SentryLevel.Debug),
+            SendClientReports = true,
+            Debug = true
+        };
+
+        var httpTransport = new HttpTransport(options, new HttpClient(httpHandler));
+
+        // some arbitrary discarded events ahead of time
+        var recorder = (ClientReportRecorder) options.ClientReportRecorder;
+        recorder.RecordDiscardedEvent(DiscardReason.BeforeSend, DataCategory.Attachment);
+        recorder.RecordDiscardedEvent(DiscardReason.EventProcessor, DataCategory.Error);
+        recorder.RecordDiscardedEvent(DiscardReason.EventProcessor, DataCategory.Error);
+        recorder.RecordDiscardedEvent(DiscardReason.QueueOverflow, DataCategory.Security);
+        recorder.RecordDiscardedEvent(DiscardReason.QueueOverflow, DataCategory.Security);
+        recorder.RecordDiscardedEvent(DiscardReason.QueueOverflow, DataCategory.Security);
+
+        // Act
+        await httpTransport.SendEnvelopeAsync(Envelope.FromEvent(new SentryEvent()));
+
+        // Assert
+        recorder.DiscardedEvents.Should().BeEquivalentTo(new Dictionary<DiscardReasonWithCategory, int>
+        {
+            // These are the original items recorded.  They should still be there.
+            {DiscardReason.BeforeSend.WithCategory(DataCategory.Attachment), 1},
+            {DiscardReason.EventProcessor.WithCategory(DataCategory.Error), 2},
+            {DiscardReason.QueueOverflow.WithCategory(DataCategory.Security), 3},
+
+            // We also expect two new items recorded, due to the forced network failure.
+            {DiscardReason.NetworkError.WithCategory(DataCategory.Error), 1},  // from the event
+            {DiscardReason.NetworkError.WithCategory(DataCategory.Default), 1} // from the client report
+        });
+    }
+
+    [Fact]
+    public async Task SendEnvelopeAsync_RateLimited_DoesNotRestoreDiscardedEventCounts()
+    {
+        // Arrange
+        using var httpHandler = new RecordingHttpMessageHandler(
+            new FakeHttpMessageHandler(
+                () => new HttpResponseMessage((HttpStatusCode)429)));
+
+        var options = new SentryOptions
+        {
+            Dsn = DsnSamples.ValidDsnWithSecret,
+            DiagnosticLogger = new TraceDiagnosticLogger(SentryLevel.Debug),
+            SendClientReports = true,
+            Debug = true
+        };
+
+        var httpTransport = new HttpTransport(options, new HttpClient(httpHandler));
+
+        // some arbitrary discarded events ahead of time
+        var recorder = (ClientReportRecorder) options.ClientReportRecorder;
+        recorder.RecordDiscardedEvent(DiscardReason.BeforeSend, DataCategory.Attachment);
+        recorder.RecordDiscardedEvent(DiscardReason.EventProcessor, DataCategory.Error);
+        recorder.RecordDiscardedEvent(DiscardReason.EventProcessor, DataCategory.Error);
+        recorder.RecordDiscardedEvent(DiscardReason.QueueOverflow, DataCategory.Security);
+        recorder.RecordDiscardedEvent(DiscardReason.QueueOverflow, DataCategory.Security);
+        recorder.RecordDiscardedEvent(DiscardReason.QueueOverflow, DataCategory.Security);
+
+        // Act
+        await httpTransport.SendEnvelopeAsync(Envelope.FromEvent(new SentryEvent()));
+
+        // Assert
+        var totalCounts = recorder.DiscardedEvents.Values.Sum();
+        Assert.Equal(0, totalCounts);
     }
 
     [Fact]
@@ -601,5 +766,84 @@ public class HttpTransportTests
 
         // Assert
         requestContent.Should().Contain(envelope.TryGetEventId().ToString());
+    }
+
+    [Fact]
+    public void ProcessEnvelope_ShouldAttachClientReport()
+    {
+        var options = new SentryOptions();
+
+        var expectedTimestamp = DateTimeOffset.MaxValue;
+        var clock = Substitute.For<ISystemClock>();
+        clock.GetUtcNow().Returns(expectedTimestamp);
+
+        var recorder = new ClientReportRecorder(options, clock);
+        options.ClientReportRecorder = recorder;
+
+        var logger = Substitute.For<IDiagnosticLogger>();
+
+        var httpTransport = Substitute.For<HttpTransportBase>(options, null, clock);
+
+        // add some fake discards for the report
+        recorder.RecordDiscardedEvent(DiscardReason.NetworkError, DataCategory.Internal);
+        recorder.RecordDiscardedEvent(DiscardReason.NetworkError, DataCategory.Security);
+        recorder.RecordDiscardedEvent(DiscardReason.QueueOverflow, DataCategory.Error);
+        recorder.RecordDiscardedEvent(DiscardReason.QueueOverflow, DataCategory.Error);
+        recorder.RecordDiscardedEvent(DiscardReason.RateLimitBackoff, DataCategory.Transaction);
+        recorder.RecordDiscardedEvent(DiscardReason.RateLimitBackoff, DataCategory.Transaction);
+        recorder.RecordDiscardedEvent(DiscardReason.RateLimitBackoff, DataCategory.Transaction);
+
+        var sentryEvent = new SentryEvent();
+        var expectedEventJson = EnvelopeItem.FromEvent(sentryEvent).Payload.SerializeToString(logger);
+
+        var envelope = Envelope.FromEvent(sentryEvent);
+        var processedEnvelope = httpTransport.ProcessEnvelope(envelope);
+
+        // There should be exactly two items in the envelope
+        Assert.Equal(2, processedEnvelope.Items.Count);
+        var eventItem = processedEnvelope.Items[0];
+        var clientReportItem = processedEnvelope.Items[1];
+
+        // Make sure they have the correct types set in their headers
+        Assert.Equal("event", eventItem.TryGetType());
+        Assert.Equal("client_report", clientReportItem.TryGetType());
+
+        // The event should be unmodified
+        Assert.Equal(expectedEventJson, eventItem.Payload.SerializeToString(logger));
+
+        // The client report should contain the counts of the fake discards
+        const string expectedClientReportJson =
+            "{\"timestamp\":\"9999-12-31T23:59:59.9999999+00:00\"," +
+            "\"discarded_events\":[" +
+            "{\"reason\":\"network_error\",\"category\":\"internal\",\"quantity\":1}," +
+            "{\"reason\":\"network_error\",\"category\":\"security\",\"quantity\":1}," +
+            "{\"reason\":\"queue_overflow\",\"category\":\"error\",\"quantity\":2}," +
+            "{\"reason\":\"ratelimit_backoff\",\"category\":\"transaction\",\"quantity\":3}" +
+            "]}";
+
+        var clientReportJson = clientReportItem.Payload.SerializeToString(logger);
+        Assert.Equal(expectedClientReportJson, clientReportJson);
+    }
+
+    [Fact]
+    public void ProcessEnvelope_ShouldNotAttachClientReportWhenOptionDisabled()
+    {
+        var options = new SentryOptions
+        {
+            // Disable sending of client reports
+            SendClientReports = false
+        };
+
+        var httpTransport = Substitute.For<HttpTransportBase>(options, null, null);
+
+        var recorder = options.ClientReportRecorder;
+        recorder.RecordDiscardedEvent(DiscardReason.QueueOverflow, DataCategory.Error);
+
+        var envelope = Envelope.FromEvent(new SentryEvent());
+        var processedEnvelope = httpTransport.ProcessEnvelope(envelope);
+
+        // There should only be the one event in the envelope
+        Assert.Equal(1, processedEnvelope.Items.Count);
+        Assert.Equal("event", processedEnvelope.Items[0].TryGetType());
     }
 }